"Hi" from branch "Interpreter"
Added a .rar file with a faked but expected return from "Interpreter".
this always returns the same and that is equal to the interpretation
expected from "put the white ball in a box on the floor".

WARNING!!!!
DO NOT LET THE OPENING OF THE .rar FILE OVERWRITE ANY FILES. Instead
copy the part that is needed into your exisiting files so that you don't
lose anything.